# gost/x/api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    AuthConfig:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    ConnectorConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    DialerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    HopConfig:
        properties:
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                type: string
                x-go-name: Interface
            name:
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/NodeConfig'
                type: array
                x-go-name: Nodes
            resolver:
                type: string
                x-go-name: Resolver
            selector:
                $ref: '#/definitions/SelectorConfig'
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    IngressConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            rules:
                items:
                    $ref: '#/definitions/IngressRuleConfig'
                type: array
                x-go-name: Rules
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    IngressRuleConfig:
        properties:
            endpoint:
                type: string
                x-go-name: Endpoint
            hostname:
                type: string
                x-go-name: Hostname
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    LimiterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            limits:
                items:
                    type: string
                type: array
                x-go-name: Limits
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    NameserverConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            chain:
                type: string
                x-go-name: Chain
            clientIP:
                type: string
                x-go-name: ClientIP
            hostname:
                type: string
                x-go-name: Hostname
            prefer:
                type: string
                x-go-name: Prefer
            timeout:
                $ref: '#/definitions/Duration'
            ttl:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    NodeConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            connector:
                $ref: '#/definitions/ConnectorConfig'
            dialer:
                $ref: '#/definitions/DialerConfig'
            host:
                type: string
                x-go-name: Host
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                type: string
                x-go-name: Interface
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            protocol:
                type: string
                x-go-name: Protocol
            resolver:
                type: string
                x-go-name: Resolver
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    ResolverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            nameservers:
                items:
                    $ref: '#/definitions/NameserverConfig'
                type: array
                x-go-name: Nameservers
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: proxy_forwarder/gost/x/api
    SelectorConfig:
        properties:
            failTimeout:
                $ref: '#/definitions/Duration'
            maxFails:
                format: int64
                type: integer
                x-go-name: MaxFails
            strategy:
                type: string
                x-go-name: Strategy
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    SockOptsConfig:
        properties:
            mark:
                format: int64
                type: integer
                x-go-name: Mark
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
...
    /config/hops:
        post:
            operationId: createHopRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/HopConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new hop, the name of hop must be unique in hop list.
            tags:
                - Hop
    /config/hops/{hop}:
        delete:
            operationId: deleteHopRequest
            parameters:
                - in: path
                  name: hop
                  required: true
                  type: string
                  x-go-name: Hop
            responses:
                "200":
                    $ref: '#/responses/deleteHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete hop by name.
            tags:
                - Hop
        put:
            operationId: updateHopRequest
            parameters:
                - in: path
                  name: hop
                  required: true
                  type: string
                  x-go-name: Hop
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/HopConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update hop by name, the hop must already exist.
            tags:
                - Hop
...
    /config/ingresses:
        post:
            operationId: createIngressRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/IngressConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createIngressResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new ingress, the name of the ingress must be unique in ingress list.
            tags:
                - Ingress
    /config/ingresses/{ingress}:
        delete:
            operationId: deleteIngressRequest
            parameters:
                - in: path
                  name: ingress
                  required: true
                  type: string
                  x-go-name: Ingress
            responses:
                "200":
                    $ref: '#/responses/deleteIngressResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete ingress by name.
            tags:
                - Ingress
        put:
            operationId: updateIngressRequest
            parameters:
                - in: path
                  name: ingress
                  required: true
                  type: string
                  x-go-name: Ingress
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/IngressConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateIngressResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update ingress by name, the ingress must already exist.
            tags:
                - Ingress
...
    /config/resolvers:
        post:
            operationId: createResolverRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ResolverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new resolver, the name of the resolver must be unique in resolver list.
            tags:
                - Resolver
    /config/resolvers/{resolver}:
        delete:
            operationId: deleteResolverRequest
            parameters:
                - in: path
                  name: resolver
                  required: true
                  type: string
                  x-go-name: Resolver
            responses:
                "200":
                    $ref: '#/responses/deleteResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete resolver by name.
            tags:
                - Resolver
        put:
            operationId: updateResolverRequest
            parameters:
                - in: path
                  name: resolver
                  required: true
                  type: string
                  x-go-name: Resolver
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ResolverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update resolver by name, the resolver must already exist.
            tags:
                - Resolver
    /config/rlimiters:
        post:
            operationId: createRateLimiterRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new rate limiter, the name of limiter must be unique in limiter list.
            tags:
                - Limiter
    /config/rlimiters/{limiter}:
        delete:
            operationId: deleteRateLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
            responses:
                "200":
                    $ref: '#/responses/deleteRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete rate limiter by name.
            tags:
                - Limiter
        put:
            operationId: updateRateLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update rate limiter by name, the limiter must already exist.
            tags:
                - Limiter
...
produces:
    - application/json
responses:
...
    createHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    createIngressResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    createRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteIngressResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateIngressResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"